Complete initialization of the board.
It is truly ugly since basically all of it is hard coded. Halfway
through though I thought of a better way, which can be implemented
later. Create a reference txt file that contains all the information for
each of the pieces, probably comma delimited so that it can be easily
split up, such as (Boardwalk,blue,200,etc.). That way I can easily split
the String line and access each of the pieces one by one.